Use new ApiSerializer extender
