+++ src/AddFlagsApiAttributes.php
@@ -1,0 +1,52 @@
+<?php
+
+/*
+ * This file is part of Flarum.
+ *
+ * For detailed copyright and license information, please view the
+ * LICENSE file that was distributed with this source code.
+ */
+
+namespace Flarum\Flags;
+
+use Flarum\Api\Serializer\ForumSerializer;
+use Flarum\Settings\SettingsRepositoryInterface;
+use Flarum\User\User;
+
+class AddFlagsApiAttributes
+{
+    /**
+     * @var SettingsRepositoryInterface
+     */
+    protected $settings;
+
+    /**
+     * @param SettingsRepositoryInterface $settings
+     */
+    public function __construct(SettingsRepositoryInterface $settings)
+    {
+        $this->settings = $settings;
+    }
+
+    public function __invoke(array $attributes, $model, ForumSerializer $serializer)
+    {
+        $attributes['canViewFlags'] = $serializer->getActor()->hasPermissionLike('discussion.viewFlags');
+
+        if ($attributes['canViewFlags']) {
+            $attributes['flagCount'] = (int) $this->getFlagCount($serializer->getActor());
+        }
+
+        $attributes['guidelinesUrl'] = $this->settings->get('flarum-flags.guidelines_url');
+
+        return $attributes;
+    }
+
+    /**
+     * @param User $actor
+     * @return int
+     */
+    protected function getFlagCount(User $actor)
+    {
+        return Flag::whereVisibleTo($actor)->distinct()->count('flags.post_id');
+    }
+}
